Master  update

+++ app/Views/admin/edit-governmental-link.php
@@ -70,7 +70,7 @@
                             <td><?= date('d-m-Y', strtotime($value['created_at'])) ?></td>
                             <td>
                                 <a href="<?= base_url() ?>admin/edit-governmental-link/<?= $value['id'] ?>" class="btn btn-sm btn-primary">Edit</a>
-                                <a href="<?= base_url() ?>admin/delete-governmental-link/<?= $value['id'] ?>" class="btn btn-sm btn-danger">Delete</a>
+                                <a href="<?= base_url() ?>admin/delete-governmental-link/<?= $value['id'] ?>" class="btn btn-sm btn-danger" onclick="return confirm('Are you sure...!')">Delete</a>
                             </td>
                         </tr>
                     <?php } ?>
